% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOpgrouptest.R
\name{IOpgrouptest}
\alias{IOpgrouptest}
\title{Tests for significance of a group of circumstances}
\usage{
IOpgrouptest(
  Y,
  X,
  FVs,
  ML = c("Lasso", "Ridge", "RF", "CIF", "XGB", "CB", "Torch", "loglin", "NLLS_exp",
    "OLSensemble", "SL"),
  iop_full = NULL,
  circs,
  OLSensemble = c("Lasso", "Ridge", "RF", "CIF", "XGB", "CB"),
  SL.library = c("SL.ranger", "SL.xgboost", "SL.glmnet"),
  ensemble = c("SL.Lasso", "SL.Ridge", "SL.RF", "SL.CIF", "SL.XGB", "SL.CB"),
  ensemblefolds = 5,
  weights = NULL,
  rf.cf.ntree = 500,
  rf.depth = 5,
  cf.depth = 5,
  polynomial.Lasso = 1,
  polynomial.Ridge = 1,
  polynomial.loglin = 1,
  xgb.nrounds = 200,
  xgb.max.depth = 6,
  cb.iterations = 1000,
  cb.depth = 6,
  torch.epochs = 50,
  torch.hidden_units = c(64, 32),
  torch.lr = 0.01,
  torch.dropout = 0.2,
  mtry = max(floor(ncol(X)/3), 1),
  polynomial.NLLS_exp = 1,
  start_nlls = NULL
)
}
\arguments{
\item{Y}{is a vector containing the (continuous) outcome of interest}

\item{X}{is a dataframe containing all the circumstances}

\item{FVs}{fitted values taken from running the function IOp first}

\item{ML}{is a string specifying which machine learner to use, usually one uses
the same ML as the one used in the estimation of IOp}

\item{iop_full}{an IOp estimate based on the Gini using the full set of circumstances}

\item{circs}{vector of circumstances for which we want to perform the group test}

\item{ensemble}{is a string vector specifying which learners
should be used in the SuperLearner}

\item{weights}{survey weights adding up to 1}
}
\value{
test results
}
\description{
\code{IOpgrouptest} takes the output of IOp estimation and a selection of
circumstances for which it performs a joint test.
}
\examples{
n <- 3000
X1 <- rnorm(n)
X2 <- rnorm(n)
Y1 <- exp(2 + 0.2*X1 + 0.1*X2 + rnorm(n,0,0.5))
Y2 <- exp(2 + 0*X1 + 0.1*X2 + rnorm(n,0,0.5))
X <- data.frame(X1,X2)
ML = "XGB"
set.seed(123)

res <- IOp(Y,X,ML = ML,sterr = TRUE, IOp_rel = TRUE,
           est_method = "Debiased", fitted_values = TRUE, CFit = TRUE,
           rf.depth = 5)

FVs <- res$FVs
iop_pe <- res$IOp[1]
res$IOp
res$IOp_rel

IOpgrouptest(Y1,X, FVs, ML = "XGB", iop_full = iop_pe, circs = c("X1"))
IOpgrouptest(Y2,X, FVs, ML = "XGB", iop_full = iop_pe, circs = c("X1"))


}
\references{
Escanciano, J. C., & Terschuur, J. R. (2022).
Debiased Semiparametric U-Statistics: Machine Learning Inference
on Inequality of Opportunity. arXiv preprint arXiv:2206.05235.

Terschuur, J. (2022). Debiased Machine Learning Inequality
of Opportunity in Europe. arXiv preprint arXiv:2212.02407.
}
