% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fvs_fvs0_folds.R
\name{plot_fvs_fvs0_folds}
\alias{plot_fvs_fvs0_folds}
\title{Plot estimated fitted values and true fitted values (if available) per
cross-fitting fold}
\usage{
plot_fvs_fvs0_folds(data_list, sign_error_list)
}
\arguments{
\item{sign_error_list}{list with a vector of sgn coincidences for each pair
for each fold, can be found as output called sgns of function IOp}

\item{fold_list_fvs}{list with a dataframe per CF fold, can be found as
output called FVsinfold of function IOp}
}
\value{
Faceted plot
}
\description{
Plot estimated fitted values and true fitted values (if available) per
cross-fitting fold
}
\references{
Escanciano, J. C., & Terschuur, J. R. (2022).
Debiased Semiparametric U-Statistics: Machine Learning Inference
on Inequality of Opportunity. arXiv preprint arXiv:2206.05235.

Terschuur, J. (2022). Debiased Machine Learning Inequality
of Opportunity in Europe. arXiv preprint arXiv:2212.02407.
}
