% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOpgrouptest.R
\name{IOpgrouptest}
\alias{IOpgrouptest}
\title{Tests for significance of a group of circumstances}
\usage{
IOpgrouptest(
  Y,
  X,
  FVs,
  ineq = c("Gini", "MLD", c("Gini", "MLD")),
  ML = c("Lasso", "Ridge", "RF", "CIF", "XGB", "CB", "SL"),
  ensemble = c("SL.Lasso", "SL.Ridge", "SL.RF", "SL.CIF", "SL.XGB", "SL.CB"),
  K = 5,
  iop_gini = NULL,
  iop_mld = NULL,
  circs,
  weights = NULL
)
}
\arguments{
\item{Y}{is a vector containing the (continuous) outcome of interest}

\item{X}{is a dataframe containing all the circumstances}

\item{FVs}{fitted values taken from running the function IOp first}

\item{ineq}{which inequality index among Gini and MLD should be chosen.
We can choose to use both.}

\item{ML}{is a string specifying which machine learner to use, usually one uses
the same ML as the one used in the estimation of IOp}

\item{ensemble}{is a string vector specifying which learners
should be used in the SuperLearner}

\item{K}{in how many parts should the data be split for cross-fitting}

\item{iop_gini}{an IOp estimate based on the Gini using the full set of circumstances}

\item{iop_mld}{an IOp estimate based on the MLD using the full set of circumstances}

\item{circs}{vector of circumstances for which we want to perform the group test}

\item{weights}{survey weights adding up to 1}
}
\value{
test results
}
\description{
\code{IOpgrouptest} takes the output of IOp estimation and a selection of
circumstances for which it performs a joint test.
}
\examples{
X <- dplyr::select(mad2019,-Y)
Y <- mad2019$Y
circs <- c("educM","educF")
iop <- IOp(Y,
           X,
           est_method = "Debiased",
           CFit = TRUE,
           ineq = c("Gini","MLD"),
           plugin_method = c("ML"),
           ML = "RF",
           sterr = FALSE,
           fitted_values = TRUE)
FVs <- iop$FVs
iop_gini <- iop$IOp["IOp","Gini"]
iop_mld <- iop$IOp["IOp","MLD"]

Gtest <- IOpgrouptest(Y,
                      X,
                      FVs = FVs,
                      ineq = c("Gini","MLD"),
                      ML = "RF",
                      K = 5,
                      iop_gini = iop_gini,
                      iop_mld = iop_mld,
                      circs = circs)
Gtest$Gini
Gtest$MLD

}
\references{
Escanciano, J. C., & Terschuur, J. R. (2022).
Debiased Semiparametric U-Statistics: Machine Learning Inference
on Inequality of Opportunity. arXiv preprint arXiv:2206.05235.
}
